#FEAT: created the console module afresh & renamed the former after francis
